requirements.txt
numpy
pandas
pyarrow
pytest
python-dotenv
requests
